:recycle: [REFACTOR] jwtTokenProvider를 사용하여 사용자 조회
