+++ src/main/java/com/outofcity/server/service/activity/ActivityFavoritiesService.java
@@ -3,8 +3,8 @@ package com.outofcity.server.service.activity;
 import com.outofcity.server.domain.Activity;
 import com.outofcity.server.domain.ActivityFavorities;
 import com.outofcity.server.domain.GeneralMember;
-import com.outofcity.server.dto.activity.ActivityFavoritiesResponseDto;
-import com.outofcity.server.dto.activity.ActivityResponseDto;
+import com.outofcity.server.dto.activity.response.ActivityFavoritiesResponseDto;
+import com.outofcity.server.dto.activity.response.ActivityResponseDto;
 import com.outofcity.server.global.exception.BusinessException;
 import com.outofcity.server.global.exception.message.ErrorMessage;
 import com.outofcity.server.global.exception.message.SuccessMessage;
@@ -18,6 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -35,7 +36,10 @@ public class ActivityFavoritiesService {
     // 액티비티 즐겨찾기 목록 조회
     public List<ActivityResponseDto> getFavorites(String token) {
         // 토큰으로 일반회원 조회
-        GeneralMember generalMember = generalMemberRepository.findByToken(token, jwtTokenProvider);
+        Long generalMemberId = jwtTokenProvider.getUserFromJwt(token);
+
+        GeneralMember generalMember = generalMemberRepository.findById(generalMemberId)
+                .orElseThrow(() -> new BusinessException(ErrorMessage.GENERAL_MEMBER_NOT_FOUND));
         // 일반회원의 액티비티 즐겨찾기 목록 조회
         List<ActivityFavorities> activityFavorities = activityFavoritiesRepository.findAllByGeneralMember(generalMember);
 
@@ -50,8 +54,12 @@ public class ActivityFavoritiesService {
 
     // 액티비티 즐겨찾기 추가
     public ActivityFavoritiesResponseDto addFavorite(String token, Long activityId) {
+
+        Long generalMemberId = jwtTokenProvider.getUserFromJwt(token);
+
         // 토큰으로 일반회원 조회
-        GeneralMember generalMember = generalMemberRepository.findByToken(token, jwtTokenProvider);
+        GeneralMember generalMember = generalMemberRepository.findById(generalMemberId)
+                .orElseThrow(() -> new BusinessException(ErrorMessage.GENERAL_MEMBER_NOT_FOUND));
         // 액티비티 조회
         Activity activity = activityRepository.findById(activityId)
                 .orElseThrow(() -> new BusinessException(ErrorMessage.ACTIVITY_NOT_FOUND));
@@ -74,8 +82,12 @@ public class ActivityFavoritiesService {
 
     // 액티비티 즐겨찾기 삭제
     public SuccessMessage deleteFavorite(String token, Long activityId) {
+
+        Long generalMemberId = jwtTokenProvider.getUserFromJwt(token);
+
         // 토큰으로 일반회원 조회
-        GeneralMember generalMember = generalMemberRepository.findByToken(token, jwtTokenProvider);
+        GeneralMember generalMember = generalMemberRepository.findById(generalMemberId)
+                .orElseThrow(() -> new BusinessException(ErrorMessage.GENERAL_MEMBER_NOT_FOUND));
 
         // 액티비티 조회
         Activity activity = activityRepository.findById(activityId)
